requirements.txt
google-generativeai==0.3.0
customtkinter==5.2.0
pyttsx3==2.90
python-dotenv==1.0.0
SpeechRecognition==3.10.0
sounddevice==0.4.6
numpy==1.26.4
Pillow==10.2.0
Pillow==10.2.0
